Begin splitting redux actions/reducers into sector, system, and planet
