+++ src/components/hex-map/index.js
@@ -2,8 +2,8 @@ import { connect } from 'react-redux';
 
 import HexMap from './hex-map';
 
-const mapStateToProps = ({ sector }) => ({
-  holdKey: sector.holdKey,
+const mapStateToProps = ({ system }) => ({
+  holdKey: system.holdKey,
 });
 
 export default connect(mapStateToProps)(HexMap);
