+++ src/components/system-tooltips/index.js
@@ -2,9 +2,9 @@ import { connect } from 'react-redux';
 
 import SystemTooltips from './system-tooltips';
 
-const mapStateToProps = ({ sector }) => ({
-  hoverKey: sector.hoverKey,
-  holdKey: sector.holdKey,
+const mapStateToProps = ({ system }) => ({
+  hoverKey: system.hoverKey,
+  holdKey: system.holdKey,
 });
 
 export default connect(mapStateToProps)(SystemTooltips);
