+++ src/store/reducers/index.js
@@ -1,5 +1,7 @@
 import sector from './sector.reducers';
+import system from './system.reducers';
 
 export default {
   sector,
+  system,
 };
